requirements.txt
python_dateutil == 2.6.0
setuptools >= 21.0.0
flask-sqlalchemy >= 2.5.0
rdflib == 6.2.0
requests_oauthlib == 1.3.0
psycopg2 >= 2.9.0
python-dotenv
openapi-spec-validator==0.4.0
hypercorn==0.14.3
fastapi[all]
python-multipart
pycurl
shieldapi
pre-commit
flake8
